requirements.txt
# Core RAG dependencies
langchain>=0.2.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.3
langchain-core>=0.2.0
langchain-ollama>=0.1.0

# Vector store
faiss-cpu>=1.7.4

# PDF loading
pypdf>=4.0.0

# Environment variable management
python-dotenv>=1.0.0

# Web UI
streamlit>=1.35.0

# Sentence embeddings (pulled transitively, pinned for reproducibility)
sentence-transformers>=2.7.0

# HTTP requests for Ollama connection checking
requests>=2.31.0
